# daemon/openapi.yaml
openapi: 3.1.0
...
info:
  title: Hazyflow
  version: "1.0.0"
  description: |
    The Hazyflow gateway. One surface for humans, scripts, and LLM
    agents — the web UI, the `hz-mcp` MCP server, and any third-party
    integrations all talk to these endpoints.

    Auth is bearer-token. Session cookies are also accepted for
    browser callers but undocumented here.

    Discoverability:
      - `GET /api/v1` — service descriptor + links to everything below.
      - `GET /api/v1/openapi.json` — this document, served live.
      - `GET /api/v1/catalog` — a one-page summary of installed
        integrations and drops, sized to fit in one LLM context.

  contact:
    name: Hazyflow
    url: https://hazy.flow
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.hazy.flow
    description: Production
  - url: /
    description: Same-origin (e.g. local dev gateway, self-hosted deployments)

# Bearer auth applies to every operation unless explicitly overridden.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/me/secrets:
    get:
      tags: [Secrets]
      operationId: listSecrets
      summary: List secret names (values never returned)
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/SecretSummary" }
  /api/v1/me/secrets/{name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string, pattern: "^[A-Za-z0-9_./-]+$" } }
    put:
      tags: [Secrets]
      operationId: putSecret
      summary: Upsert a secret value
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: string }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "204": { description: upserted }
    delete:
      tags: [Secrets]
      operationId: deleteSecret
      summary: Delete a secret
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "204": { description: deleted }

  # ---------- Connections (OAuth) ------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: hzk_*

  # ---------- Common parameters --------------------------------------
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 128 }
      description: |
        Opaque client-chosen string. Re-submitting the same request
        with the same key within 24h returns the original response.
...
  responses:
...
    Unauthorized:
      description: No or invalid credential.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: "Machine-readable enum, snake_case."
            message:
              type: string
              description: Human/LLM-readable explanation.
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
            doc:
              type: string
              description: Optional deep link into this OpenAPI doc for the offending schema.
...
    SecretSummary:
      type: object
      required: [name]
      properties:
        name:       { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...